+++ main.c
@@ -23,9 +23,31 @@
 #include <stdbool.h>
 #include "logging.h"
 #include "nfc_utils.h"
+#include "commands.c"
 
+int main(int argc, char *argv[], char *envp[]){
 
-int main() {
+  //Defult options
+  set_skip_confirmation(false);
+  set_eeprom_size(SRIX4K_EEPROM_SIZE);
+  set_eeprom_blocks_amount(SRIX4K_EEPROM_BLOCKS);
+  set_verbose(false);
+
+  // Parse options
+  int opt = 0;
+  while ((opt = getopt(argc, argv, "hvyt:")) != -1) {
+      switch (opt) {
+          case 'v': set_verbose(true); break;
+          case 'y':set_skip_confirmation(true); break;
+          case 't':
+              if (strcmp(optarg, "512") == 0) {
+                  set_eeprom_size(SRI512_EEPROM_SIZE);
+                  set_eeprom_blocks_amount(SRI512_EEPROM_BLOCKS);
+              }
+              break;
+          
+      }
+  }
 
   int choice = 0;
 
@@ -41,7 +63,7 @@ int main() {
         printf(GREEN "6) " RESET "Modify block manually\n" );
         printf(GREEN "7) " RESET "Write EEPROM file to NFC tag\n" );
         printf(GREEN "8) " RESET "Reset OTP Blocks\n" );
-        printf(GREEN "9) " RESET "Modify config file\n" );
+        printf(GREEN "9) " RESET "Help\n" );
         printf(GREEN "0) " RESET "Exit\n" );  
 
         printf(YELLOW "\n>>> Choose an option: " RESET);
@@ -49,14 +71,14 @@ int main() {
 
         switch (choice){
             case 1: system("nfc-list"); break;
-            case 2: system("./commands/read_eeprom_content"); break;  
-            case 3: system("./commands/read_tag_info"); break;
-            case 4: system("./commands/write_eeprom_to_file"); break;
-            case 5: system("./commands/read_eeprom_file"); break;
-            case 6: system("./commands/modify_block"); break;
-            case 7: system("./commands/write_to_tag"); break;
-            case 8: system("./commands/otp_reset"); break;
-            case 9: system("nano ./config"); break;        
+            case 2: read_eeprom_content(); break;  
+            case 3: read_tag_info(); break;
+            case 4: write_eeprom_to_file(); break;
+            case 5: read_eeprom_file(); break;
+            case 6: modfiy_block();break;
+            case 7: write_to_tag(); break;
+            case 8: otp_reset(); break;
+            case 9: print_options(argv[0]); break;        
             case 0: exit(0);
         }