+++ CHANGELOG.md
@@ -1,4 +1,21 @@
 # Changelog
 
-## v1.0.0
-* First public release (December 18, 2022)
+## v1.2.0 (December 21, 2022)
+
+* Improvements & fix issues 
+* Added `help` command
+* Added `commands` file
+* Renamed `nfc-srix-programmer` command to `nfc-srix`
+* Removed `commands` Commands Folder
+* Removed `config` Config file
+* Removed `modify_block` file
+* Removed `otp_reset` file
+* Removed `read_eeprom_content` file
+* Removed `read_eeprom_file` file
+* Removed `read_tag_info` file
+* Removed `write_eeprom_to_file` file
+* Removed `write_to_tag` file
+
+## v1.0.0 (December 18, 2022)
+
+* First public release